Change order of clientSecret and clientId
While it make makes perfect sense to put the id first and the secret second, the https://auth.nello.io/admin/ has the order reversed. A naive user (like me) would just copy and paste them and assume that the order is the same both on the API-page and in the default example. So: surrender to the API-page which can _not_ be changed with a pull request 😉